chore: move provider and models into llm directory
Move packages/server/src/provider/* and packages/server/src/models/*
into packages/server/src/llm/provider/ to consolidate all LLM-related
code under a single directory in preparation for new provider types.

Update server routes from /provider and /models to /llm/provider and
/llm/models, and update all web client fetch calls accordingly.
