+++ packages/server/src/index.ts
@@ -12,9 +12,9 @@ import { configRouter } from '@/routes/config.js';
 import { connectorsRouter } from '@/routes/connectors.js';
 import { eventsRouter } from '@/routes/events.js';
 import { mcpRouter } from '@/routes/mcp.js';
-import { modelsRouter } from '@/routes/models.js';
+import { modelsRouter } from '@/routes/llm-models.js';
 import { permissionsRouter } from '@/routes/permissions.js';
-import { providerRouter } from '@/routes/provider.js';
+import { providerRouter } from '@/routes/llm-provider.js';
 import { questionsRouter } from '@/routes/questions.js';
 import { queueRouter } from '@/routes/queue.js';
 import { settingsRouter } from '@/routes/settings.js';
@@ -56,8 +56,8 @@ app.route('/chat', queueRouter);
 app.route('/config', configRouter);
 app.route('/events', eventsRouter);
 app.route('/mcp', mcpRouter);
-app.route('/models', modelsRouter);
-app.route('/provider', providerRouter);
+app.route('/llm/models', modelsRouter);
+app.route('/llm/provider', providerRouter);
 app.route('/settings', settingsRouter);
 app.route('/shortcuts', shortcutsRouter);
 app.route('/usage', usageRouter);
