+++ packages/server/src/llm/attachment-transform.ts
@@ -1,4 +1,4 @@
-import * as Models from '@/provider/models.js';
+import * as Models from '@/llm/provider/models.js';
 import type { ModelMessage, FilePart, ImagePart, TextPart } from 'ai';
 
 type Modality = 'image' | 'video' | 'audio' | 'pdf';
