+++ packages/server/src/llm/provider/model-visibility.ts
