+++ packages/server/src/llm/provider/service.ts
@@ -4,10 +4,10 @@ import { getDb } from '@/db/client.js';
 import { providerConfig } from '@/db/schema.js';
 import { err, isServiceError, ok } from '@/lib/service-result.js';
 import type { ServiceResult } from '@/lib/service-result.js';
-import * as ProviderLogos from '@/provider/logos.js';
-import { isAllowedProvider } from '@/provider/models.js';
-import * as Models from '@/provider/models.js';
-import { ProviderCredentialsSchema } from '@/provider/provider.js';
+import * as ProviderLogos from '@/llm/provider/logos.js';
+import { isAllowedProvider } from '@/llm/provider/models.js';
+import * as Models from '@/llm/provider/models.js';
+import { ProviderCredentialsSchema } from '@/llm/provider/provider.js';
 
 type ProviderSummary = {
   id: string;
