+++ packages/server/src/llm/stream/runner.ts
@@ -25,8 +25,8 @@ import {
   setSessionActiveToolsetIds,
 } from '@/llm/stream/session-toolsets.js';
 import { executeStepWithRetry, type StepOptions } from '@/llm/stream/step-executor.js';
-import { createProvider } from '@/provider/provider.js';
-import type { ProviderCredentials } from '@/provider/provider.js';
+import { createProvider } from '@/llm/provider/provider.js';
+import type { ProviderCredentials } from '@/llm/provider/provider.js';
 import { createTaskTool } from '@/tools/core/task.js';
 import { createToolsetTools } from '@/tools/core/toolset-management.js';
 import { createTools, MAX_STEPS, MAX_STEPS_WARNING } from '@/tools/runtime/registry.js';
