+++ packages/server/src/routes/llm-models.ts
@@ -3,7 +3,7 @@ import { Hono } from 'hono';
 import { z } from 'zod';
 
 import { isServiceError } from '@/lib/service-result.js';
-import { deleteVisibility, listVisibilityOverrides, upsertVisibility } from '@/models/service.js';
+import { deleteVisibility, listVisibilityOverrides, upsertVisibility } from '@/llm/provider/model-visibility.js';
 
 export const modelsRouter = new Hono();
